requirements2.txt
numpy<2.0.0
pandas>=2.1.0
requests>=2.31.0
pytesseract>=0.3.10
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
